+++ cosid-core/src/main/java/me/ahoo/cosid/snowflake/SecondSnowflakeId.java
@@ -29,19 +29,19 @@ public class SecondSnowflakeId extends AbstractSnowflakeId {
     public static final int DEFAULT_SEQUENCE_BIT = 22;
     public static final long DEFAULT_SEQUENCE_RESET_THRESHOLD = ~(-1L << (DEFAULT_SEQUENCE_BIT - 1));
     
-    public SecondSnowflakeId(long machineId) {
+    public SecondSnowflakeId(int machineId) {
         this(CosId.COSID_EPOCH_SECOND, DEFAULT_TIMESTAMP_BIT, DEFAULT_MACHINE_BIT, DEFAULT_SEQUENCE_BIT, machineId, DEFAULT_SEQUENCE_RESET_THRESHOLD);
     }
     
-    public SecondSnowflakeId(int machineBit, long machineId) {
+    public SecondSnowflakeId(int machineBit, int machineId) {
         super(CosId.COSID_EPOCH_SECOND, DEFAULT_TIMESTAMP_BIT, machineBit, DEFAULT_SEQUENCE_BIT, machineId, DEFAULT_SEQUENCE_RESET_THRESHOLD);
     }
     
-    public SecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, long machineId) {
+    public SecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, int machineId) {
         super(epoch, timestampBit, machineBit, sequenceBit, machineId, SnowflakeId.defaultSequenceResetThreshold(sequenceBit));
     }
     
-    public SecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, long machineId, long sequenceResetThreshold) {
+    public SecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, int machineId, long sequenceResetThreshold) {
         super(epoch, timestampBit, machineBit, sequenceBit, machineId, sequenceResetThreshold);
     }
